♻️ refactoring: merge post/user ssr fetcher module
