+++ src/app/components/activity.component.tsx
@@ -5,6 +5,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { swrHttp } from "../utility/fetcher";
 import { PostCard } from "./card.component";
 import { Alert } from "./alert.component";
+import { UserStore } from "../store/user";
 import { Utility } from "../utility";
 import { useState } from "react";
 import useSWR from "swr";
@@ -13,6 +14,7 @@ const utility = new Utility();
 
 export function UserActivity(props: IUserActivityProps) {
   const { id, sort } = props;
+  const { user_id } = UserStore();
   const [page, setPage] = useState<number>(1);
   const { data, isLoading, error } = useSWR<IPaginationData<IPostPreview[]>>(
     {
@@ -46,7 +48,7 @@ export function UserActivity(props: IUserActivityProps) {
                 isPublic={true}
                 profileImage={e.profile_image}
                 createdAt={e.created_at}
-                buttonLink={`/post/${e.id}`}
+                buttonLink={user_id ? `/post/myPost/${e.id}` : `/post/${e.id}`}
                 key={i}
               />
             );
