+++ src/app/layout/toast.layout.tsx
@@ -1,6 +1,6 @@
 "use client";
-import React, { ReactNode } from "react";
 import { Toaster } from "react-hot-toast";
+import type { ReactNode } from "react";
 
 export function ToastLayout({ children }: { children: ReactNode }) {
   return (
