+++ src/app/post/myPost/[id]/page.tsx
@@ -65,11 +65,7 @@ export default function ViewPrivatePostPage({ params }: IPostProps) {
               <span className="text-gray-100 text-base">{post.author}</span>
             </Link>
             <p className="text-gray-500 text-base mt-2">{dayjs(post.created_at).format("YYYY.MM.DD HH:mm:ss")} 게시됨</p>
-            {post.edited_at && post.edited_at !== 0 ? (
-              <p className="text-gray-500 text-base">{dayjs(post.edited_at).format("YYYY.MM.DD HH:mm:ss")} 수정됨</p>
-            ) : (
-              <></>
-            )}
+            {post.edited_at !== 0 && <p className="text-gray-500 text-base">{dayjs(post.edited_at).format("YYYY.MM.DD HH:mm:ss")} 수정됨</p>}
             <PostOption postId={post.id} />
             <hr className="border-gray-700 my-5" />
             <pre className="pretendard text-wrap overflow-hidden">
