+++ src/app/post/myPost/layout.tsx
@@ -1,5 +1,5 @@
 import AuthLayout from "@/app/layout/auth.layout";
-import { ReactNode } from "react";
+import type { ReactNode } from "react";
 
 export default function PrivatePostLayout({ children }: { children: ReactNode }) {
   return <AuthLayout>{children}</AuthLayout>;
