+++ src/app/utility/modal.ts
@@ -1,5 +1,5 @@
+import type { IConfirmModalProps } from "../types";
 import { confirmAlert } from "react-confirm-alert";
-import { IConfirmModalProps } from "../types";
 
 export function confirmModal(props: IConfirmModalProps) {
   const { title, message, callback } = props;
