+++ src/app/write/layout.tsx
@@ -1,6 +1,6 @@
 import { ReCaptchaProvider } from "next-recaptcha-v3";
 import AuthLayout from "../layout/auth.layout";
-import { ReactNode } from "react";
+import type { ReactNode } from "react";
 
 export default function WriteLayout({ children }: { children: ReactNode }) {
   return (
